Revert "Enable lock screen rotation"
This reverts commit 08ee97543d86bc935734a07bf034354f123ff0cc.

This way made it hard for advanced users to change the behaviour.
A better way to enable the lock screen rotation is to use
lockscreen.rot_override property. Setting it to 'true' will enable
the rotation while users will still be able to change it to 'false'
in build.prop if they need it disabled.

Also, it will become obsolete anyway after the merge of:
http://review.cyanogenmod.org/44378
http://review.cyanogenmod.org/44379