gps: add function to enable/disable GPS engine lock
From http://review.cyanogenmod.org/29531

Engine lock is enabled on loc_eng_init, disabled on loc_eng_cleanup.

The engine lock state is non-volatile, and control of the lock is needed
to ensure that the GPS works and is responsive to requests.  Otherwise the
GPS may silently cease to work depending on its last state in stock firmware.

Change-Id: I7936ddfaa97983ad0533b809f7b2f064d1c783e0